confirmacion de next y prev pag

+++ src/paginas/Favoritos.pagina.jsx
@@ -14,10 +14,10 @@ const PaginaFavoritos = () => {
 const {favoritesId, favorites} = useAppSelector ((state) => state.characters);
 const dispatch = useAppDispatch ();
 useEffect ( () => {
-dispatch(loadCharactersFavorites());
+dispatch(loadCharactersFavorites())
 },[favoritesId, dispatch]);
-console.log(favoritesId)
-    return <div className="container">
+if (favorites.length === 0 ) return <div className="container"> No tienes Favoritos</div>
+return <div className="container">
         <div className="actions">
             <h3>Personajes Favoritos</h3>
             <button className="danger" onClick={()=> dispatch(charactersSlice.actions.deleteAllFavorites())}>Eliminar Favoritos</button>
